requirements.txt
Flask==2.2.2

PyMySQL~=1.0.2
DBUtils~=3.0.2

Pillow~=9.2.0

schedule~=1.1.0

PyJWT~=2.4.0
